ci: add Cloudflare Pages PR previews
